Rename package logic to loadbalancer and include groupsreader in loadbalancer
