+++ dfaasagent/agent/loadbalancer/hacfgtypes.go
@@ -1,4 +1,4 @@
-package logic
+package loadbalancer
 
 import (
 	"time"
