+++ dfaasagent/agent/loadbalancer/haproxycfgnms.tmpl
