+++ dfaasagent/agent/loadbalancer/loadbalancer.go
@@ -1,4 +1,4 @@
-package logic
+package loadbalancer
 
 import (
 	"sync"
